[refactor] extract save logic to utils
